+++ locust/main.py
@@ -13,9 +13,9 @@ import random
 from optparse import OptionParser
 
 import web
-import inspectlocust
-from log import setup_logging
-from locust.stats import stats_printer, RequestStats, print_percentile_stats, print_error_report, print_stats
+from log import setup_logging, console_logger
+from stats import stats_printer, RequestStats, print_percentile_stats, print_error_report, print_stats
+from inspectlocust import print_task_ratio, get_task_ratio_dict
 from core import Locust, WebLocust
 from runners import MasterLocustRunner, SlaveLocustRunner, LocalLocustRunner
 
@@ -188,12 +188,13 @@ def parse_options():
         default=False,
         help="print table of the locust classes' task execution ratio"
     )
+    # Display ratio table of all tasks in JSON format
     parser.add_option(
-        '--show-task-ratio-confluence',
+        '--show-task-ratio-json',
         action='store_true',
-        dest='show_task_ratio_confluence',
+        dest='show_task_ratio_json',
         default=False,
-        help="print the locust classes' task execution ratio in confluence list markup"
+        help="print json data of the locust classes' task execution ratio"
     )
     
     # Version number (optparse gives you --version but we have to do it
@@ -354,21 +355,20 @@ def run():
         locust_classes = locusts.values()
     
     if options.show_task_ratio:
-        print "\n Task ratio per locust class"
-        print "-" * 80
-        inspectlocust.print_task_ratio(locust_classes)
-        print "\n Total task ratio"
-        print "-" * 80
-        inspectlocust.print_task_ratio(locust_classes, total=True)
+        console_logger.info("\n Task ratio per locust class")
+        console_logger.info( "-" * 80)
+        print_task_ratio(locust_classes)
+        console_logger.info("\n Total task ratio")
+        console_logger.info("-" * 80)
+        print_task_ratio(locust_classes, total=True)
         sys.exit(0)
-    
-    if options.show_task_ratio_confluence:
-        print "\nh1. Task ratio per locust class"
-        print
-        inspectlocust.print_task_ratio_confluence(locust_classes)
-        print "\nh1. Total task ratio"
-        print
-        inspectlocust.print_task_ratio_confluence(locust_classes, total=True)
+    if options.show_task_ratio_json:
+        from json import dumps
+        task_data = {
+            "per_class": get_task_ratio_dict(locust_classes), 
+            "total": get_task_ratio_dict(locust_classes, total=True)
+        }
+        console_logger.info(dumps(task_data))
         sys.exit(0)
     
     # if --master is set, make sure --no-web isn't set