+++ locust/stats.py
@@ -3,19 +3,9 @@ import gevent
 from copy import copy
 import math
 
-from exception import InterruptLocust
 import events
-
-# set up logger for the statistics tables
-import logging
-stats_logger = logging.getLogger("stats_logger")
-# create console handler
-sh = logging.StreamHandler()
-sh.setLevel(logging.INFO)
-# formatter that doesn't include anything but the message
-sh.setFormatter(logging.Formatter('%(message)s'))
-stats_logger.addHandler(sh)
-stats_logger.propagate = False
+from exception import InterruptLocust
+from log import console_logger
 
 STATS_NAME_WIDTH = 60
 
@@ -30,7 +20,8 @@ class RequestStats(object):
     global_start_time = None
     errors = {}
 
-    def __init__(self, name):
+    def __init__(self, method, name):
+        self.method = method
         self.name = name
         self.num_reqs_per_sec = {}
         self.last_request_timestamp = 0
@@ -241,16 +232,16 @@ class RequestStats(object):
         )
 
     @classmethod
-    def get(cls, name):
-        request = cls.requests.get(name, None)
+    def get(cls, method, name):
+        request = cls.requests.get((method, name), None)
         if not request:
-            request = RequestStats(name)
-            cls.requests[name] = request
+            request = RequestStats(method, name)
+            cls.requests[(method, name)] = request
         return request
 
     @classmethod
     def sum_stats(cls, name="Total", full_request_history=False):
-        stats = RequestStats(name)
+        stats = RequestStats(None, name)
         for s in cls.requests.itervalues():
             stats.iadd_stats(s, full_request_history)
         return stats
@@ -291,15 +282,15 @@ def percentile(N, percent, key=lambda x:x):
     d1 = key(N[int(c)]) * (k-f)
     return d0+d1
 
-def on_request_success(name, response_time, response):
+def on_request_success(method, name, response_time, response):
     if RequestStats.global_max_requests is not None and RequestStats.total_num_requests >= RequestStats.global_max_requests:
         raise InterruptLocust("Maximum number of requests reached")
 
     content_length = int(response.info.getheader("Content-Length") or 0)
-    RequestStats.get(name).log(response_time, content_length)
+    RequestStats.get(method, name).log(response_time, content_length)
 
-def on_request_failure(name, response_time, error, response=None):
-    RequestStats.get(name).log_error(error)
+def on_request_failure(method, name, response_time, error, response=None):
+    RequestStats.get(method, name).log_error(error)
 
 def on_report_to_master(client_id, data):
     data["stats"] = [RequestStats.requests[name].get_stripped_report() for name in RequestStats.requests if not (RequestStats.requests[name].num_reqs == 0 and RequestStats.requests[name].num_failures == 0)]
@@ -323,8 +314,8 @@ events.slave_report += on_slave_report
 
 
 def print_stats(stats):
-    stats_logger.info((" %-" + str(STATS_NAME_WIDTH) + "s %7s %12s %7s %7s %7s  | %7s %7s") % ('Name', '# reqs', '# fails', 'Avg', 'Min', 'Max', 'Median', 'req/s'))
-    stats_logger.info("-" * (80 + STATS_NAME_WIDTH))
+    console_logger.info((" %-" + str(STATS_NAME_WIDTH) + "s %7s %12s %7s %7s %7s  | %7s %7s") % ('Name', '# reqs', '# fails', 'Avg', 'Min', 'Max', 'Median', 'req/s'))
+    console_logger.info("-" * (80 + STATS_NAME_WIDTH))
     total_rps = 0
     total_reqs = 0
     total_failures = 0
@@ -333,31 +324,31 @@ def print_stats(stats):
         total_rps += r.current_rps
         total_reqs += r.num_reqs
         total_failures += r.num_failures
-        stats_logger.info(r)
-    stats_logger.info("-" * (80 + STATS_NAME_WIDTH))
+        console_logger.info(r)
+    console_logger.info("-" * (80 + STATS_NAME_WIDTH))
 
     try:
         fail_percent = (total_failures/float(total_reqs))*100
     except ZeroDivisionError:
         fail_percent = 0
 
-    stats_logger.info((" %-" + str(STATS_NAME_WIDTH) + "s %7d %12s %42.2f") % ('Total', total_reqs, "%d(%.2f%%)" % (total_failures, fail_percent), total_rps))
-    stats_logger.info("")
+    console_logger.info((" %-" + str(STATS_NAME_WIDTH) + "s %7d %12s %42.2f") % ('Total', total_reqs, "%d(%.2f%%)" % (total_failures, fail_percent), total_rps))
+    console_logger.info("")
 
 def print_percentile_stats(stats):
-    stats_logger.info("Percentage of the requests completed within given times")
-    stats_logger.info((" %-" + str(STATS_NAME_WIDTH) + "s %8s %6s %6s %6s %6s %6s %6s %6s %6s %6s") % ('Name', '# reqs', '50%', '66%', '75%', '80%', '90%', '95%', '98%', '99%', '100%'))
-    stats_logger.info("-" * (80 + STATS_NAME_WIDTH))
+    console_logger.info("Percentage of the requests completed within given times")
+    console_logger.info((" %-" + str(STATS_NAME_WIDTH) + "s %8s %6s %6s %6s %6s %6s %6s %6s %6s %6s") % ('Name', '# reqs', '50%', '66%', '75%', '80%', '90%', '95%', '98%', '99%', '100%'))
+    console_logger.info("-" * (80 + STATS_NAME_WIDTH))
     complete_list = []
     for key in sorted(stats.iterkeys()):
         r = stats[key]
         if r.response_times:
-            stats_logger.info(r.percentile())
+            console_logger.info(r.percentile())
             complete_list.extend(r.create_response_times_list())
-    stats_logger.info("-" * (80 + STATS_NAME_WIDTH))
+    console_logger.info("-" * (80 + STATS_NAME_WIDTH))
     complete_list.sort()
     if complete_list:
-        stats_logger.info( (" %-" + str(STATS_NAME_WIDTH) + "s %8s %6d %6d %6d %6d %6d %6d %6d %6d %6d") % (
+        console_logger.info( (" %-" + str(STATS_NAME_WIDTH) + "s %8s %6d %6d %6d %6d %6d %6d %6d %6d %6d") % (
             'Total',
             str(len(complete_list)),
             percentile(complete_list, 0.5),
@@ -370,16 +361,16 @@ def print_percentile_stats(stats):
             percentile(complete_list, 0.99),
             complete_list[-1]
         ))
-    stats_logger.info("")
+    console_logger.info("")
 
 def print_error_report():
-    stats_logger.info("Error report")
-    stats_logger.info(" %-18s %-100s" % ("# occurences", "Error"))
-    stats_logger.info("-" * (80 + STATS_NAME_WIDTH))
+    console_logger.info("Error report")
+    console_logger.info(" %-18s %-100s" % ("# occurences", "Error"))
+    console_logger.info("-" * (80 + STATS_NAME_WIDTH))
     for error, count in RequestStats.errors.iteritems():
-        stats_logger.info(" %-18i %-100s" % (count, error))
-    stats_logger.info("-" * (80 + STATS_NAME_WIDTH))
-    stats_logger.info("")
+        console_logger.info(" %-18i %-100s" % (count, error))
+    console_logger.info("-" * (80 + STATS_NAME_WIDTH))
+    console_logger.info("")
 
 def stats_printer():
     from runners import locust_runner