+++ locust/test/test_stats.py
@@ -3,12 +3,13 @@ import time
 
 from testcases import WebserverTestCase
 from locust.stats import RequestStats
-from locust.core import Locust
+from locust.core import Locust, SubLocust, task
+from locust.inspectlocust import get_task_ratio_dict
 
 class TestRequestStats(unittest.TestCase):
     def setUp(self):
         RequestStats.global_start_time = time.time()
-        self.s = RequestStats("test_entry")
+        self.s = RequestStats("GET", "test_entry")
         self.s.log(45, 0)
         self.s.log(135, 0)
         self.s.log(44, 0)
@@ -53,13 +54,13 @@ class TestRequestStats(unittest.TestCase):
         self.assertEqual(self.s.median_response_time, 85)
 
     def test_aggregation(self):
-        s1 = RequestStats("aggregate me!")
+        s1 = RequestStats("GET", "aggregate me!")
         s1.log(12, 0)
         s1.log(12, 0)
         s1.log(38, 0)
         s1.log_error("Dummy exzeption")
 
-        s2 = RequestStats("aggregate me!")
+        s2 = RequestStats("GET", "aggregate me!")
         s2.log_error("Dummy exzeption")
         s2.log_error("Dummy exzeption")
         s2.log(12, 0)
@@ -70,7 +71,7 @@ class TestRequestStats(unittest.TestCase):
         s2.log(55, 0)
         s2.log(97, 0)
 
-        s = RequestStats("")
+        s = RequestStats("GET", "")
         s.iadd_stats(s1, full_request_history=True)
         s.iadd_stats(s2, full_request_history=True)
 
@@ -86,6 +87,39 @@ class TestRequestStatsWithWebserver(WebserverTestCase):
     
         locust = MyLocust()
         locust.client.get("/ultra_fast")
-        self.assertEqual(RequestStats.get("/ultra_fast").avg_content_length, len("This is an ultra fast response"))
+        self.assertEqual(RequestStats.get("GET", "/ultra_fast").avg_content_length, len("This is an ultra fast response"))
         locust.client.get("/ultra_fast")
-        self.assertEqual(RequestStats.get("/ultra_fast").avg_content_length, len("This is an ultra fast response"))
+        self.assertEqual(RequestStats.get("GET", "/ultra_fast").avg_content_length, len("This is an ultra fast response"))
+
+
+class MyLocust(Locust):
+    @task(75)
+    def root_task(self):
+        pass
+    
+    @task(25)
+    class MySubLocust(SubLocust):
+        @task
+        def task1(self):
+            pass
+        @task
+        def task2(self):
+            pass
+    
+class TestInspectLocust(unittest.TestCase):
+    def test_get_task_ratio_dict_relative(self):
+        ratio = get_task_ratio_dict([MyLocust])
+        self.assertEqual(1.0, ratio["MyLocust"]["ratio"])
+        self.assertEqual(0.75, ratio["MyLocust"]["tasks"]["root_task"]["ratio"])
+        self.assertEqual(0.25, ratio["MyLocust"]["tasks"]["MySubLocust"]["ratio"])
+        self.assertEqual(0.5, ratio["MyLocust"]["tasks"]["MySubLocust"]["tasks"]["task1"]["ratio"])
+        self.assertEqual(0.5, ratio["MyLocust"]["tasks"]["MySubLocust"]["tasks"]["task2"]["ratio"])
+    
+    def test_get_task_ratio_dict_total(self):
+        ratio = get_task_ratio_dict([MyLocust], total=True)
+        self.assertEqual(1.0, ratio["MyLocust"]["ratio"])
+        self.assertEqual(0.75, ratio["MyLocust"]["tasks"]["root_task"]["ratio"])
+        self.assertEqual(0.25, ratio["MyLocust"]["tasks"]["MySubLocust"]["ratio"])
+        self.assertEqual(0.125, ratio["MyLocust"]["tasks"]["MySubLocust"]["tasks"]["task1"]["ratio"])
+        self.assertEqual(0.125, ratio["MyLocust"]["tasks"]["MySubLocust"]["tasks"]["task2"]["ratio"])
+