+++ docs/changelog.rst
@@ -14,6 +14,13 @@ API changes
 
 * Web inteface is now turned on by default. The **--web** command line option has been replaced by --no-web.
 
+Improvements and bug fixes
+--------------------------
+
+* Removed **--show-task-ratio-confluence** and added a **--show-task-ratio-json** option instead. The
+  **--show-task-ratio-json** will output JSON data containing the task execution ratio for the locust
+  "brain".
+
 0.4
 ===